Use custom class loader instead of Spring-Boot PropertiesLauncher
Rationale: PropertiesLauncher works very well when starting the application
with 'java -jar', but it does not work well when starting the application
with 'gradle bootRun'  or from IDE, because in those cases the main class
will be started directly.

To accomodate for same behavior on DEV machines as on e.g. PROD environment,
a custom classloader, checking for jars in a configured directory, has been implemented.